Reorganized exports and imports so circular dependencies work properly
